% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bats.R
\name{BATS}
\alias{BATS}
\title{BATS model}
\usage{
BATS(formula, ...)
}
\arguments{
\item{formula}{A TBATS model formula (see details).}

\item{...}{Additional arguments (see details).}
}
\value{
A TBATS model, analogous to other model objects within fable/fabletools.
}
\description{
A \link{fable} wrapper for \link{forecast}[bats]
}
\details{
Accepts and parses several model specials.
}
\note{
Maybe some other day.
}
