% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbats.R
\name{TBATS}
\alias{TBATS}
\title{TBATS model}
\usage{
TBATS(formula, ...)
}
\arguments{
\item{formula}{A TBATS model formula (see details).}

\item{...}{Additional arguments (see details).}
}
\value{
A TBATS model, analogous to other model objects within fable/fabletools.
}
\description{
A \link{fable} wrapper for \link{forecast}[tbats]
}
\details{
Accepts and parses several model specials.
}
\note{
Maybe some other day.
}
